apps-standalone/fa-de-onboarding-foundations/requirements.txt
black==25.1.0
click==8.1.8
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.7
pyaml==25.1.0
PyYAML==6.0.2
